Move AddImport and related functions to a separate file.
